nothing

+++ app/models.py
@@ -1,9 +1,9 @@
 from pydantic import BaseModel
 
 class SensorData(BaseModel):
-    temperature: float
-    humidity: float
-    moisture: float
+    temperature: int
+    humidity: int
+    moisture: int
 
 class TaskCreate(BaseModel):
     title: str
